Language Configurations
This was largely inspired by snbt-vscode, which provided a lot of inspiration for making this work. I plan on combining this with something like vscode-nbt, so you essentially have a plain-text way of editing NBT using SNBT in the editor, and you will get type checking thanks to an LSP service driven by the TypeScript definitions I'm going to make for all of the game's internal types. The NBT <-> SNBT conversion for the editing process is going to be driven by NBTify under the hood, and it will handle the file saving and format specs. I want to pipe NBTify's format options configurations into a menu available to the user, possibly through the Command Palette. Then you would simply just choose things like the file endian type and compression using the Command Palette.

I wanted to start making Dovetail better recently, but I also feel like trying this out first will get me further in the long run, and close to something actually really useful and ready for people to make things with. Then I will have to figure out how to add editing support for things like MCA files, and maybe other things like GAMEDATA and Bedrock MCWORLD files. I was going to do that all in Dovetail, or some other web-first app, but I think I probably should do that as the second iteration of building all of this, as trying to both get it all working, *and* getting that to work on the web and everywhere, it may take some of the strength out of getting the project started. And from the things I've been learning recently, one of them is to just try and get something started at all. It doesn't have to be perfect, because getting it started at all is better than not doing anything because you want it to be perfect.

https://github.com/misode/vscode-nbt
https://github.com/Tnze/snbt-vscode