Lexing + JSON-Inspired Lang Config
Started reworking the scope selectors for the tmLanguage definition file, but then I looked into it more and found out that I think it worlds similarly to CSS classes (that's probably what it's build with, too?), so you can have a 'standard' TM identifier, followed by a language-specific one, which is what themes can use to hook into specific languages individually. Really neat! I was going to remove all of the NBT-specific identifiers from the selectors, but then I found in the definition files for JSON that it does the same thing, and it turns out it's probably why vscode-snbt does it this way too :)

https://macromates.com/manual/en/language_grammars#naming-conventions
https://www.apeth.com/nonblog/stories/textmatebundle.html
https://github.com/microsoft/vscode/blob/main/extensions/json/syntaxes/JSON.tmLanguage.json
https://github.com/microsoft/vscode/blob/main/extensions/json/language-configuration.json

Found all of these last night, but it was getting kind of late, so I held off adding these until I had fresh eyes for it the next day.

This one was a very great find! It's an example for a custom programming language lexer/parser, and not only was it helpful for understanding things here like for syntax highlighting, but I also realized that it really helps explain nice ways to handle SNBT reading and writing too! Rather than using tokens and using the parts of the lexing process to use as program instructions, I instead will use them as NBT primitives on the JavaScript side. That's what I'm doing now basically, I just hadn't thought of the crossover like that before.
https://www.youtube.com/watch?v=8VB5TY1sIRo

Kind of unrelated here, but I found a possible nice way to get Minecraft Bedrock running on Linux, it's with a project called Waydroid! The Linux launcher is good and all, but it relies on having a custom wrapper fork maintained to run the game, while this only relies on Bedrock being supported on Android, which is something that we essentially have come to expect being safe, I'd imagine lol. Not sure if it's feasible, but I could probably hack this together somehow to get that running on Android, in Linux, emulated on macOS too. Sounds kind of bad though, but never know, could be worth the try!
https://www.reddit.com/r/linux_gaming/comments/uhojw7/minecraft_bedrock_via_waydroid_on_amd_and_intel/